docs: fix outdated comments in .importlinter
There were references to itemstore, which hasn't existed in a while. I
also reorganized the comments to be inline with the apps in the Core App
Dependency Layering section, since this section is going to grow over
time, and it's easier to understand if the commments are right next to
the app paths.