feat: admin helper for finding third-party 1:1 models
Created the one_to_one_related_model_html helper to let us link 1:1
models from lower-level models to higher-level ones. So for instance,
"publishing" is a lower level app than "components", meaning that
components can call into publishing, but publishing should have no
awareness of components. But it's really convenient to be able to click
on a PublishableEntity and go to the corresponding Component.

This is even more important as people add more models, potentially
even in edx-platform plugins.