feat: convert mime_type field to its own model.
This commit normalizes the Media/MIME type in RawContent to be a foriegn
key to a new table rather than a text string. This was done for a couple
of reasons:

1. Recent problems with edx-platform's courseware_studentmodule table
   have driven home the point that we should try to reduce index sizes
   where possible, and RawContent is likely to grow into a very large
   table in time.
2. Media type designations can sometimes change, and that's much easier
   to do as a data migration if it's in its own lookup table, rather
   than affecting millions of rows in RawContent.

In order to support these changes, I've also added LRU caching for
MediaType lookup... which in turn required a test helper that will clear
out the values between test calls. That's the rationale behind our new
openedx_learning.lib.test_utils.TestCase.