feat: add features to support v2 libraries
This commit adds a number of features needed to accomodate content
libraries. It also tunes a number of queries to elminated n+1 fetch
issues when tested with libraries, by using select_related more
liberally, and following cached relationships instead of re-fetching
things from the database.

As part of this general tuning, PublishableEntityVersion.version_num was
reduced to a 4-byte PositiveIntegerField instead of using the 8-byte
PositiveBigIntegerField. This should save us a bit of space with our
indexes while still allowing for over two billion versions for any given
PublishableEntity.