fix: map PublishableEntity keys to library conventions
PublishableEntities have their own "key" field that must be unique
within a LearningPackage. Components are PublishableEntities (via 1:1
relationship), but they store that data as separate fields that are
useful for filtering and grouping on. When we create a Component, it
creates a PublishableEntity key value based on its own namespace, type,
and local key.

Before this commit, we followed a more ModuleStore-like convention and
the derived key would have its type and local key separated by "@":

  xblock.v1:html@424337d3-e1fc-4970-822a-684b1ccb0c1d

This commit changes it so that we follow the v2 libraries convention of
having a ":" between the two instead:

  xblock.v1:html:424337d3-e1fc-4970-822a-684b1ccb0c1d