refactor!: move openedx_learning.core -> openedx_learning.apps.authoring
All the Django apps that were in openedx_learning.core were specific to
authoring library and course content: publishing, contents, components.
As time goes on, we will likely add other groupings of apps, e.g.
"learner", "personalization", "grading", etc.

Important compatibility notes:

1. This is a breaking change, and edx-platform will need to be updated
   to point to the new locations for apps.
2. This should not affect the table names or migrations,  since all app
   names remain the same, and we use the explicit repo-level "oel_"
   prefix anyway.

This commit also removes the rest_api app. This app was never really
functional, and was just put in as a skeleton to demonstrate a possible
approach to adding a REST API to the openedx_learning package as a
distinct app (as opposed to making a separate API per app). That being
said, there's no current use case to expose a REST API from any of the
authoring apps, and it's not clear whether we'd really want a top level
rest_api app or separate rest_api apps for different groups of apps
(e.g. "the REST API for authoring"). To avoid confusion, I'm just
removing it altogether.