test: update test_units_containing to catch bad JOINs in the query
Both before and after this change, test_units_containing() was working correctly, but without this change in place, it's easy to refactor test_units_containing(..., ignore_pinned=True) in a way that still passes this test but is actually incorrect.

e.g. the following implementation is wrong, but was passing the test without this fix in place:
```
    qs = Container.objects.filter(
        publishable_entity__draft__version__containerversion__entity_list__entitylistrow__entity_id=publishable_entity_pk,
    ).order_by("pk")  # Ordering is mostly for consistent test cases.
    if ignore_pinned:
        qs = qs.filter(
            publishable_entity__draft__version__containerversion__entity_list__entitylistrow__entity_version_id=None,
        )
```