feat: track publishing dependencies and side-effects
Create the PublishSideEffect model as the publishing analog for
the DraftSideEffect model, and create these side effects for
parent containers when one of the children is published. This
allows us to efficiently query when any subtree of content was
affected by a publish.

Also introduces PublishableEntityVersionDependency as a way of
tracking unpinned dependencies of a version, e.g. the children
of a version of a Unit.

Also introduces a new dependencies_hash_digest field to Draft
and Published models to track dependency state.