feat: Browsable Django Admin interface for Containers (#330)
This adds several admin pages which were previously missing or barebones:
Section, Subsection, Unit, Container, and EntityList. The Container and
EntityList pages have inlines tables to display their related ContainerVersions
and EntityListsRows. This enables developers/operators to browse up and down
the content hierarchies that are stored in Learning Core.