refactor: remove UUID field from ComponentVersionContent
It's wasteful to have a UUID on this join table, as we already have
a unique way to address everything in it by the unique combination
of the ComponentVersion (PublishableEntityVersion) and the
ComponentVersionContent.key (file path).

Dropping this column may cause transient errors for some sites if
someone is creating components after the column was dropped, but
before the new app code is live. UUIDs are generated on the Python
side, so Django will try to generate a UUID to put into a column
that has already been dropped on the database side.

That being said, this window of time should be small, and this
feature is not in widespread use yet. I don't think a more complex
migration is warranted.