feat: publishing dependencies and state tracking (#369)
This commit introduces the idea of dependencies to the publishing app.
Dependencies are a generalization of unpinned container children that
we plan to later use for things like course files & uploads, grading
policies, and other things where a change in one publishable entity
affects the state of another publishable entity, even if they do not
share a direct parent/child relationship.

Dependencies do not *replace* container children, but they offer a
simpler model that allows the publishing app to quickly calculate
things like, "Do any of the transitive dependencies of this entity
have unpublished changes?". It also means that we can efficiently
implement "publish this entity and all its dependencies at the same
time" without having to use any callbacks to other apps that might
implement different kinds of dependency relationships (e.g. grading
policy). As long as those other apps declare their dependencies when
calling create_publishable_entity_version(), the publishing app will
have enough information to do these types of queries.

It's worth noting that we never really wanted to put container logic
directly in the publishing app, but we were forced to because the
process of publishing is so coupled to parent/child relationships.
This shift to thinking about dependencies rather than children will
make it possible for us shift container logic out of publishing and
into its own app.

This commit also creates the PublishSideEffect as the publishing
equivalent to DraftSideEffect. We've come to the tentative decision
that we should probably try to unify some of these parallel draft/
published models, but that will be a bigger undertaking.

This commit also introduces a dependencies_hash_digest field to
DraftChangeLogRecord and PublishLogRecord in order to track the state
of all live (current draft/published) dependencies. Tracking this has
allowed us to optimize contains_unpublished_changes() to be a single
query.

Finally, this commit adds Django admin functionality for this new
data. This admin interface is still rudimentary though, and could use
a lot of work.
---------

Co-authored-by: Kyle McCormick <kyle@kylemccormick.me>